chore: spans only log events in verbose mode
